pyproject.toml
[project]
name = "ai-automation-framework"
version = "0.1.0"
description = "AI-Driven Automation Infrastructure with Agentic AI, RAG, and MCP"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    # Core Testing Framework
    "pytest>=8.4.0",
    "pytest-playwright>=0.7.0",
    # Environment & Configuration
    "python-dotenv>=1.0.0",
    # MCP Server
    "fastmcp>=2.12.0",
    # AI & RAG - Core
    "langchain-core>=1.0.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-ollama>=1.0.0",
    "langchain-chroma>=0.1.0",
    "chromadb>=1.1.0",
    # AI & RAG - Future Use (Google Gemini integration)
    "langchain-google-genai>=2.0.0",
    "google-generativeai>=0.8.0",
    "google-genai>=1.45.0",
    # Data Validation
    "pydantic>=2.0.0",
    "allure-pytest>=2.15.3",
    "assertpy>=1.1",
    "pytest-xdist>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.0.0",
    "mypy>=1.13.0",
    "types-requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "mcp_server", "rag_engine", "tests"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
"--screenshot", "only-on-failure",
"--base-url", "https://www.saucedemo.com"
]
markers = [
    "sanity: Fast smoke tests for PR validation",
    "test_case_key: mark a test with a test case key."
]

[dependency-groups]
dev = [
    "types-assertpy>=1.1.0.20251219",
]
